+++ benchmarks/Datasets/LabeledBench.php
@@ -74,8 +74,8 @@ class LabeledBench
      * @Iterations(5)
      * @OutputTimeUnit("milliseconds", precision=3)
      */
-    public function splitByColumn() : void
+    public function splitByFeature() : void
     {
-        $this->dataset->splitByColumn(2, 3.0);
+        $this->dataset->splitByFeature(2, 3.0);
     }
 }